[runtime] Move intrinsic heap item definitions into their own file, splitting from constructor file
