+++ src/js/runtime/gc/heap_item.rs
@@ -27,28 +27,27 @@ use crate::{
         heap_item_descriptor::HeapItemDescriptor,
         interned_strings::InternedStringsSet,
         intrinsics::{
-            array_buffer_constructor::ArrayBufferObject,
+            array_buffer_object::ArrayBufferObject,
             array_iterator::ArrayIterator,
-            async_from_sync_iterator_prototype::AsyncFromSyncIterator,
-            bigint_constructor::BigIntObject,
-            boolean_constructor::BooleanObject,
-            data_view_constructor::DataViewObject,
+            async_from_sync_iterator_object::AsyncFromSyncIterator,
+            bigint_object::BigIntObject,
+            boolean_object::BooleanObject,
+            data_view_object::DataViewObject,
             date_object::DateObject,
-            error_constructor::ErrorObject,
+            error_object::ErrorObject,
             finalization_registry_object::{FinalizationRegistryCells, FinalizationRegistryObject},
-            iterator_constructor::WrappedValidIterator,
             iterator_helper_object::IteratorHelperObject,
             map_iterator::MapIterator,
             map_object::{MapObject, ValueIndexMap},
-            number_constructor::NumberObject,
+            number_object::NumberObject,
             object_prototype::ObjectPrototype,
             raw_json_object::RawJSONObject,
-            regexp_constructor::RegExpObject,
+            regexp_object::RegExpObject,
             regexp_string_iterator::RegExpStringIterator,
             set_iterator::SetIterator,
             set_object::{SetObject, ValueIndexSet},
             string_iterator::StringIterator,
-            symbol_constructor::SymbolObject,
+            symbol_object::SymbolObject,
             temporal::{
                 duration_object::DurationObject, instant_object::InstantObject,
                 plain_date_object::PlainDateObject, plain_date_time_object::PlainDateTimeObject,
@@ -61,8 +60,9 @@ use crate::{
                 Int16Array, Int32Array, UInt8Array, UInt8ClampedArray, UInt16Array, UInt32Array,
             },
             weak_map_object::{WeakMapObject, WeakValueMap},
-            weak_ref_constructor::WeakRefObject,
+            weak_ref_object::WeakRefObject,
             weak_set_object::{WeakSetObject, WeakValueSet},
+            wrapped_valid_iterator_object::WrappedValidIterator,
         },
         module::{
             import_attributes::ImportAttributes,
