+++ src/js/runtime/intrinsics/boolean_constructor.rs
@@ -1,83 +1,18 @@
-use std::mem::size_of;
-
 use crate::{
-    extend_object,
     runtime::{
-        Context, HeapItemKind, HeapPtr,
+        Context,
         alloc_error::AllocResult,
-        eval_result::EvalResult,
-        gc::{Handle, HeapItem, HeapVisitor},
+        gc::Handle,
         intrinsic_builder::IntrinsicBuilder,
-        intrinsics::{intrinsics::Intrinsic, rust_runtime::RuntimeFunction},
-        object_value::ObjectValue,
-        ordinary_object::{
-            object_create, object_create_from_constructor, object_create_with_proto,
+        intrinsics::{
+            boolean_object::BooleanObject, intrinsics::Intrinsic, rust_runtime::RuntimeFunction,
         },
+        object_value::ObjectValue,
         realm::Realm,
         type_utilities::to_boolean,
     },
-    runtime_fn, set_uninit,
+    runtime_fn,
 };
-
-// Boolean Objects (https://tc39.es/ecma262/#sec-boolean-objects)
-extend_object! {
-    pub struct BooleanObject {
-        // The boolean value wrapped by this object
-        boolean_data: bool,
-    }
-}
-
-impl BooleanObject {
-    pub fn new(cx: Context, boolean_data: bool) -> AllocResult<Handle<BooleanObject>> {
-        let mut object = object_create::<BooleanObject>(
-            cx,
-            HeapItemKind::BooleanObject,
-            Intrinsic::BooleanPrototype,
-        )?;
-
-        set_uninit!(object.boolean_data, boolean_data);
-
-        Ok(object.to_handle())
-    }
-
-    pub fn new_from_constructor(
-        cx: Context,
-        constructor: Handle<ObjectValue>,
-        boolean_data: bool,
-    ) -> EvalResult<Handle<BooleanObject>> {
-        let mut object = object_create_from_constructor::<BooleanObject>(
-            cx,
-            constructor,
-            HeapItemKind::BooleanObject,
-            Intrinsic::BooleanPrototype,
-        )?;
-
-        set_uninit!(object.boolean_data, boolean_data);
-
-        Ok(object.to_handle())
-    }
-
-    pub fn new_with_proto(
-        cx: Context,
-        proto: Handle<ObjectValue>,
-        boolean_data: bool,
-    ) -> AllocResult<Handle<BooleanObject>> {
-        let mut object =
-            object_create_with_proto::<BooleanObject>(cx, HeapItemKind::BooleanObject, proto)?;
-
-        set_uninit!(object.boolean_data, boolean_data);
-
-        Ok(object.to_handle())
-    }
-
-    pub fn boolean_data(&self) -> bool {
-        self.boolean_data
-    }
-
-    pub fn set_boolean_data(&mut self, boolean_data: bool) {
-        self.boolean_data = boolean_data;
-    }
-}
 
 pub struct BooleanConstructor;
 
@@ -110,14 +45,4 @@ impl BooleanConstructor {
             }
         }
     }}
-}
-
-impl HeapItem for BooleanObject {
-    fn byte_size(_: HeapPtr<Self>) -> usize {
-        size_of::<BooleanObject>()
-    }
-
-    fn visit_pointers(boolean_object: HeapPtr<Self>, visitor: &mut impl HeapVisitor) {
-        boolean_object.visit_object_pointers(visitor);
-    }
 }
