+++ src/js/runtime/intrinsics/iterator_constructor.rs
@@ -1,24 +1,24 @@
 use crate::{
-    extend_object, intrinsic_methods,
+    intrinsic_methods,
     runtime::{
-        Context, HeapItemKind, HeapPtr, Value,
+        Context, HeapItemKind, Value,
         abstract_operations::{call, call_object, get_method, ordinary_has_instance},
         alloc_error::AllocResult,
         collections::array::ValueArray,
         error::type_error,
         eval_result::EvalResult,
-        gc::{Handle, HeapItem, HeapVisitor},
+        gc::Handle,
         intrinsic_builder::IntrinsicBuilder,
         intrinsics::{
             intrinsics::Intrinsic, iterator_helper_object::IteratorHelperObject,
-            rust_runtime::RuntimeFunction,
+            rust_runtime::RuntimeFunction, wrapped_valid_iterator_object::WrappedValidIterator,
         },
         iterator::{create_iter_result_object, get_iterator_flattenable},
         object_value::ObjectValue,
-        ordinary_object::{object_create_from_constructor, object_create_with_proto},
+        ordinary_object::object_create_from_constructor,
         realm::Realm,
     },
-    runtime_fn, set_uninit,
+    runtime_fn,
 };
 
 pub struct IteratorConstructor;
@@ -108,53 +108,6 @@ impl IteratorConstructor {
 
         Ok(WrappedValidIterator::new(cx, iterator.iterator, iterator.next_method)?.as_value())
     }}
-}
-
-// A WrappedValidIterator wraps an iterator object and its next method.
-extend_object! {
-    pub struct WrappedValidIterator {
-        iterator: HeapPtr<ObjectValue>,
-        next_method: Value,
-    }
-}
-
-impl WrappedValidIterator {
-    fn new(
-        cx: Context,
-        iterator: Handle<ObjectValue>,
-        next_method: Handle<Value>,
-    ) -> AllocResult<Handle<ObjectValue>> {
-        let mut object = object_create_with_proto::<WrappedValidIterator>(
-            cx,
-            HeapItemKind::WrappedValidIterator,
-            cx.get_intrinsic(Intrinsic::WrapForValidIteratorPrototype),
-        )?;
-
-        set_uninit!(object.iterator, *iterator);
-        set_uninit!(object.next_method, *next_method);
-
-        Ok(object.as_object().to_handle())
-    }
-
-    fn iterator(&self) -> Handle<ObjectValue> {
-        self.iterator.to_handle()
-    }
-
-    fn next_method(&self, cx: Context) -> Handle<Value> {
-        self.next_method.to_handle(cx)
-    }
-}
-
-impl HeapItem for WrappedValidIterator {
-    fn byte_size(_: HeapPtr<Self>) -> usize {
-        size_of::<WrappedValidIterator>()
-    }
-
-    fn visit_pointers(mut wrapped_valid_iterator: HeapPtr<Self>, visitor: &mut impl HeapVisitor) {
-        wrapped_valid_iterator.visit_object_pointers(visitor);
-        visitor.visit_pointer(&mut wrapped_valid_iterator.iterator);
-        visitor.visit_value(&mut wrapped_valid_iterator.next_method);
-    }
 }
 
 /// The WrapForValidIteratorPrototype Object (https://tc39.es/ecma262/#sec-%wrapforvaliditeratorprototype%-object)
