+++ src/js/runtime/intrinsics/mod.rs
@@ -1,11 +1,13 @@
 pub mod aggregate_error_constructor;
 mod aggregate_error_prototype;
 pub mod array_buffer_constructor;
+pub mod array_buffer_object;
 mod array_buffer_prototype;
 pub mod array_constructor;
 pub mod array_from_async_generator;
 pub mod array_iterator;
 mod array_prototype;
+pub mod async_from_sync_iterator_object;
 pub mod async_from_sync_iterator_prototype;
 mod async_function_constructor;
 mod async_function_prototype;
@@ -14,16 +16,20 @@ mod async_generator_function_prototype;
 pub mod async_generator_prototype;
 mod async_iterator_prototype;
 pub mod bigint_constructor;
+pub mod bigint_object;
 mod bigint_prototype;
 pub mod boolean_constructor;
+pub mod boolean_object;
 mod boolean_prototype;
 pub mod data_view_constructor;
+pub mod data_view_object;
 mod data_view_prototype;
 mod date_constructor;
 pub mod date_object;
 pub mod date_prototype;
 mod encodings;
 pub mod error_constructor;
+pub mod error_object;
 pub mod error_prototype;
 mod finalization_registry_constructor;
 pub mod finalization_registry_object;
@@ -50,6 +56,7 @@ mod map_prototype;
 mod math_object;
 pub mod native_error;
 pub mod number_constructor;
+pub mod number_object;
 mod number_prototype;
 mod object_constructor;
 pub mod object_prototype;
@@ -59,6 +66,7 @@ pub mod proxy_constructor;
 pub mod raw_json_object;
 mod reflect_object;
 pub mod regexp_constructor;
+pub mod regexp_object;
 pub mod regexp_prototype;
 pub mod regexp_string_iterator;
 pub mod rust_runtime;
@@ -70,17 +78,21 @@ pub mod string_constructor;
 pub mod string_iterator;
 pub mod string_prototype;
 pub mod symbol_constructor;
+pub mod symbol_object;
 mod symbol_prototype;
 pub mod temporal;
 pub mod typed_array;
 mod typed_array_constructor;
+mod typed_array_object;
 mod typed_array_prototype;
 mod utils;
 mod weak_map_constructor;
 pub mod weak_map_object;
 mod weak_map_prototype;
 pub mod weak_ref_constructor;
+pub mod weak_ref_object;
 mod weak_ref_prototype;
 mod weak_set_constructor;
 pub mod weak_set_object;
 mod weak_set_prototype;
+pub mod wrapped_valid_iterator_object;
