+++ src/js/runtime/intrinsics/symbol_constructor.rs
@@ -1,52 +1,19 @@
-use std::mem::size_of;
-
 use crate::{
-    extend_object, intrinsic_methods,
+    intrinsic_methods,
     runtime::{
-        Context, Handle, HeapItemKind, HeapPtr,
+        Context, Handle,
         alloc_error::AllocResult,
         collections::hash_map::BsHashMapField,
         error::type_error,
-        gc::{HeapItem, HeapVisitor},
         intrinsic_builder::IntrinsicBuilder,
         intrinsics::{intrinsics::Intrinsic, rust_runtime::RuntimeFunction},
         object_value::ObjectValue,
-        ordinary_object::object_create,
         realm::Realm,
         type_utilities::to_string,
         value::SymbolValue,
     },
-    runtime_fn, set_uninit,
+    runtime_fn,
 };
-
-// Symbol Objects (https://tc39.es/ecma262/#sec-symbol-objects)
-extend_object! {
-    pub struct SymbolObject {
-        // The symbol value wrapped by this object
-        symbol_data: HeapPtr<SymbolValue>,
-    }
-}
-
-impl SymbolObject {
-    pub fn new_from_value(
-        cx: Context,
-        symbol_data: Handle<SymbolValue>,
-    ) -> AllocResult<Handle<SymbolObject>> {
-        let mut object = object_create::<SymbolObject>(
-            cx,
-            HeapItemKind::SymbolObject,
-            Intrinsic::SymbolPrototype,
-        )?;
-
-        set_uninit!(object.symbol_data, *symbol_data);
-
-        Ok(object.to_handle())
-    }
-
-    pub fn symbol_data(&self) -> Handle<SymbolValue> {
-        self.symbol_data.to_handle()
-    }
-}
 
 pub struct SymbolConstructor;
 
@@ -143,15 +110,4 @@ impl SymbolConstructor {
 
         Ok(cx.undefined())
     }}
-}
-
-impl HeapItem for SymbolObject {
-    fn byte_size(_: HeapPtr<Self>) -> usize {
-        size_of::<SymbolObject>()
-    }
-
-    fn visit_pointers(mut symbol_object: HeapPtr<Self>, visitor: &mut impl HeapVisitor) {
-        symbol_object.visit_object_pointers(visitor);
-        visitor.visit_pointer(&mut symbol_object.symbol_data);
-    }
 }
