+++ src/js/runtime/intrinsics/weak_ref_constructor.rs
@@ -1,67 +1,17 @@
-use std::mem::size_of;
-
 use crate::{
-    extend_object,
     runtime::{
-        Context, Handle, HeapItemKind, HeapPtr, Value,
+        Context, Handle, Value,
         alloc_error::AllocResult,
         error::type_error,
-        eval_result::EvalResult,
-        gc::{HeapItem, HeapVisitor},
         intrinsic_builder::IntrinsicBuilder,
-        intrinsics::{intrinsics::Intrinsic, rust_runtime::RuntimeFunction},
+        intrinsics::{
+            intrinsics::Intrinsic, rust_runtime::RuntimeFunction, weak_ref_object::WeakRefObject,
+        },
         object_value::ObjectValue,
-        ordinary_object::object_create_from_constructor,
         realm::Realm,
     },
-    runtime_fn, set_uninit,
+    runtime_fn,
 };
-
-// WeakRef Objects (https://tc39.es/ecma262/#sec-weak-ref-objects)
-extend_object! {
-    pub struct WeakRefObject {
-        // Weakly held reference to a value. Can only be an object, symbol, or undefined.
-        weak_ref_target: Value,
-        // Holds the address of the next weak ref that has been visited during garbage collection.
-        // Unused outside of garbage collection.
-        next_weak_ref: Option<HeapPtr<WeakRefObject>>,
-    }
-}
-
-impl WeakRefObject {
-    pub fn new_from_constructor(
-        cx: Context,
-        constructor: Handle<ObjectValue>,
-        value: Handle<Value>,
-    ) -> EvalResult<Handle<WeakRefObject>> {
-        let mut object = object_create_from_constructor::<WeakRefObject>(
-            cx,
-            constructor,
-            HeapItemKind::WeakRefObject,
-            Intrinsic::WeakRefPrototype,
-        )?;
-
-        set_uninit!(object.weak_ref_target, *value);
-
-        Ok(object.to_handle())
-    }
-
-    pub fn weak_ref_target(&self) -> Value {
-        self.weak_ref_target
-    }
-
-    pub fn set_weak_ref_target(&mut self, weak_ref_target: Value) {
-        self.weak_ref_target = weak_ref_target;
-    }
-
-    pub fn next_weak_ref(&self) -> Option<HeapPtr<WeakRefObject>> {
-        self.next_weak_ref
-    }
-
-    pub fn set_next_weak_ref(&mut self, next_weak_ref: Option<HeapPtr<WeakRefObject>>) {
-        self.next_weak_ref = next_weak_ref;
-    }
-}
 
 pub struct WeakRefConstructor;
 
@@ -116,18 +66,5 @@ pub fn can_be_held_weakly(cx: Context, value: Value) -> bool {
             .any(|(_, symbol)| symbol.ptr_eq(&symbol_value))
     } else {
         false
-    }
-}
-
-impl HeapItem for WeakRefObject {
-    fn byte_size(_: HeapPtr<Self>) -> usize {
-        size_of::<WeakRefObject>()
-    }
-
-    fn visit_pointers(mut weak_ref_object: HeapPtr<Self>, visitor: &mut impl HeapVisitor) {
-        weak_ref_object.visit_object_pointers(visitor);
-        visitor.visit_weak_value(&mut weak_ref_object.weak_ref_target);
-
-        // Intentionally do not visit next_weak_ref
     }
 }
